Moved jobs and users up in the directory structure, removed client dir
